MGMT-4314 KubeAPI: InstallEnv support for nmstate config
Adding NMState config support in KubeAPI.
To link between an InstallEnv to NMState (either one or more):
1. InstallEnv CR: add a label (currently named adi.io.my.domain/selector-nmstate-config-name)
   to nmStateConfigLabelSelector with a user defined value.
2. NMState CR: Specify the same label + value in Object metadata.

Upon InstallEnv creation, the InstallEnv controller will search by label+value for matching
NMState resources and construct a config to be sent as StaticNetworkConfig as a part
of ImageCreateParams. The backend does all validations, and currently, there is no handling
of configuration conflicts (e.g., two nmstate resources using the same MAC address).

The InstallEnv controller will watch for NMState config creation/changes and search
for corresponding InstallEnv resources to reconcile since we need to regenerate the
image for those.

Examples (soon to be migrated to assisted-service repo) may be found here:
https://github.com/nmagnezi/assisted-service-crs/blob/master/installEnv.yaml
https://github.com/nmagnezi/assisted-service-crs/blob/master/nmstate.yaml